Remove serialisation from the Crux core
This changes how effects work in the crux core to allow Rust clients
to access the Core over a fully typed API with no serialisation involved.

The main change is that Steps and Effects are stacked differently. Previously,
Effects were the payload of Step. The Step (which is now an enum depending on
its semantics - None, Once or Many) is now a payload of Effect instead, and
is generic over a capability Operation. This means Steps themselves can be
resolved, and the get wrapped in Effect to facilitate dispatch to the
right capability implementation in the Shell.

This also necessitated some changes to how CapabilitContext gets constructed
and specialised.

Left to do is an implementation of the bridge, which will contain the
serialisation code and step registry.