Docs: Suggestions from following the Getting Started guide
I've added a few pointers for things I got stuck on or had to check when following the 'Getting Started' guide in the book.
Some of these will be helpful for Rust devs that aren't that familiar with app development, and others hopefully avoid getting stuck where there's sensitivity to exact naming or setup.