android examples: compileDebugKotlin > preBuild
I found that if you was to clean build from gradle (android studio) it would break as the
typeGen and bindGen tasks would run after the copiler checks if all the
kotlin code exists.

Moving the tasks to be required by preBuild instead of
compileDebugKotlin fixed the issue

To be clear this is a very minor issue as when coding in android you want
to have already ran these tasks for code complete to work as intended.
However maybe not a minor issue if you build in CI