crux_core/command: Keep Command pending if a context is alive
If a future using a CommandContext is stored (e.g. for caching)
but the associated Command terminates before, crux will panic when
using the context due to the inner channel being disconnected since
the Command (holding the Receiver) was dropped.

Another implementation would be to look at the try_recv error and
mark the stream Ready when both reports the Disconnected error,
but this cannot be done today as the Command holds a CommandContext
to be able to provide the `spawn` api.