bridge: remove `erased_serde` dependency and usage
`erased_serde` is adding a good amount of bloat to the end binary, and
prevent us to improve the registry, like the ability to resolve with
typed responses.

In order to remove it, re-use and unify the concept of `FfiFormat` from
the middlware module and make `ResolveSerialized` generic over it to
allow for generic deserialization without erasing the types.

This is both a breaking change and regresssion as it remove the ability
to pass custom `Deserializer` (e.g. serde_json::Value) in favor of just
bytes. The later use case feeling niche enough to be removed.